Started refactoring of the exporter interface
The currently embedded exporters for .owl and .dot files are to be extracted into separate classes, using a unified interface (thus, allowing for further expansion of the exporting functionality). The exporter configuration and integration mechanisms are now there, but do not yet export data. The .dot exporter is on it's way and is almost done (only the string generation and file saving functionality is missing). Furthermore, the logging functionality was moved into a separate lib to keep things organized.